Update: Code maintenance and Deployment

+++ src/main/java/com/project/SWP391/responses/dto/LaundryDetailInfoDTO.java
@@ -7,8 +7,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.Set;
-
 @Data
 @Builder
 @AllArgsConstructor
@@ -16,8 +14,8 @@ import java.util.Set;
 public class LaundryDetailInfoDTO {
     private Long id;
     private float price;
-    private int from;
-    private int to;
+    private float from;
+    private float to;
     private String unit;
 
     @JsonIgnore
